+++ lib/src/main/kotlin/com/maltaisn/recurpicker/format/RRuleFormatter.kt
@@ -22,6 +22,8 @@ import com.maltaisn.recurpicker.Recurrence.Period
 import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.Locale
+import java.util.TimeZone
+import kotlin.contracts.contract
 
 /**
  * Utility class to write a [Recurrence] as a RRule and read it back.
@@ -34,6 +36,14 @@ import java.util.Locale
 class RRuleFormatter {
 
     /**
+     * The timezone used for formatting and parsing end dates (`UNTIL` attribute),
+     * unless the end date uses UTC time (time form #2, section 3.3.5).
+     */
+    var timeZone: TimeZone = TimeZone.getDefault()
+
+    private val dateFormat = SimpleDateFormat("", Locale.ROOT)
+
+    /**
      * Parse a RFC 5545 string recurrence rule and return a recurrence.
      * Note that this method is designed to parse only the output of [format].
      *
@@ -42,20 +52,20 @@ class RRuleFormatter {
      * yearly but not on the same day as start date, this information is lost when parsing,
      * since yearly recurrence can only happen on the same day as start date.
      *
-     * @throws IllegalArgumentException If recurrence rule cannot be parsed.
+     * @throws RRuleParseException If recurrence rule cannot be parsed.
      */
     fun parse(rrule: String): Recurrence {
-        require(rrule.startsWith(RRULE_SIGNATURE)) { "Recurrence rule string is invalid." }
+        parseError(rrule.startsWith(RRULE_SIGNATURE)) { "Recurrence rule string is invalid." }
 
         val attrs = rrule.substring(RRULE_SIGNATURE.length).split(';').associate {
             val pos = it.indexOf('=')
-            it.substring(0, pos) to it.substring(pos + 1)
+            it.substring(0, pos).toUpperCase(Locale.ROOT) to it.substring(pos + 1)
         }
 
         val period = parsePeriod(attrs)
 
-        return try {
-            Recurrence(period) {
+        return Recurrence(period) {
+            try {
                 frequency = attrs["INTERVAL"]?.toInt() ?: 1
                 if (period == Period.WEEKLY) {
                     parseWeeklyDetails(attrs)
@@ -63,21 +73,22 @@ class RRuleFormatter {
                     parseMonthlyDetails(attrs)
                 }
                 parseEndTypeDetails(attrs)
+            } catch (e: NumberFormatException) {
+                parseError("Bad number format in recurrence rule.", e)
             }
-        } catch (e: NumberFormatException) {
-            throw IllegalArgumentException("Bad number format in recurrence rule.", e)
         }
     }
 
     private fun parsePeriod(attrs: Map<String, String>): Period {
-        val periodStr = requireNotNull(attrs["FREQ"]) { "Recurrence rule must specify period." }
+        val periodStr = attrs["FREQ"]
+        parseError(periodStr != null) { "Recurrence rule must specify period." }
         return when (periodStr) {
             "NONE" -> Period.NONE
             "DAILY" -> Period.DAILY
             "WEEKLY" -> Period.WEEKLY
             "MONTHLY" -> Period.MONTHLY
             "YEARLY" -> Period.YEARLY
-            else -> throw IllegalArgumentException("Unsupported recurrence period.") // Secondly, minutely, hourly
+            else -> parseError("Unsupported recurrence period.") // Secondly, minutely, hourly
         }
     }
 
@@ -88,7 +99,7 @@ class RRuleFormatter {
         if (daysAttr != null) {
             for (dayStr in daysAttr.split(',')) {
                 val index = BYDAY_VALUES.indexOf(dayStr)
-                require(index >= 0) { "Invalid day of week literal." }
+                parseError(index >= 0) { "Invalid day of week literal." }
                 days = days or (1 shl (index + 1))
             }
             setDaysOfWeek(days)
@@ -100,7 +111,7 @@ class RRuleFormatter {
         val byDay = attrs["BYDAY"]
         if (byDay != null) {
             val day = BYDAY_VALUES.indexOf(byDay.takeLast(2))
-            require(day >= 0) { "Invalid day of week literal." }
+            parseError(day >= 0) { "Invalid day of week literal." }
             val week = attrs["BYSETPOS"]?.toInt() ?: byDay.dropLast(2).toInt()
             setDayOfWeekInMonth(1 shl (day + 1), week)
         } else {
@@ -111,15 +122,26 @@ class RRuleFormatter {
     private fun Recurrence.Builder.parseEndTypeDetails(attrs: Map<String, String>) {
         val endDateStr = attrs["UNTIL"]
         if (endDateStr != null) {
-            endDate = requireNotNull(DATE_FORMAT.parse(endDateStr)) {
-                "Invalid end date format '$endDateStr'."
-            }.time
+            endDate = parseDate(endDateStr)
         } else {
             val endCountStr = attrs["COUNT"]
             if (endCountStr != null) {
                 endCount = endCountStr.toInt()
             }
         }
+    }
+
+    private fun parseDate(dateStr: String): Long {
+        for (pattern in DATE_PATTERNS) {
+            if (dateStr.length == pattern.length) {
+                if (dateFormat.toPattern() != pattern.pattern) {
+                    dateFormat.applyPattern(pattern.pattern)
+                }
+                dateFormat.timeZone = pattern.timeZone ?: timeZone
+                return dateFormat.parse(dateStr)?.time ?: continue
+            }
+        }
+        parseError("Invalid date format '$dateStr'.")
     }
 
     /**
@@ -197,7 +219,11 @@ class RRuleFormatter {
             EndType.NEVER -> Unit
             EndType.BY_DATE -> {
                 sb.append("UNTIL=")
-                sb.append(DATE_FORMAT.format(r.endDate))
+                if (dateFormat.toPattern() != FORMAT_DATE_PATTERN) {
+                    dateFormat.applyPattern(FORMAT_DATE_PATTERN)
+                }
+                dateFormat.timeZone = timeZone
+                sb.append(dateFormat.format(r.endDate))
                 sb.append(';')
             }
             EndType.BY_COUNT -> {
@@ -208,10 +234,36 @@ class RRuleFormatter {
         }
     }
 
-    companion object {
-        private val BYDAY_VALUES = arrayOf("SU", "MO", "TU", "WE", "TH", "FR", "SA")
-        private val DATE_FORMAT = SimpleDateFormat("yyyyMMdd'T'HHmmss", Locale.ENGLISH)
+    private data class DatePattern(val pattern: String, val timeZone: TimeZone?, val length: Int)
 
+    companion object {
         private const val RRULE_SIGNATURE = "RRULE:"
+
+        private val DATE_PATTERNS = listOf(
+            DatePattern("yyyyMMdd", null, 8),
+            DatePattern("yyyyMMdd'T'HHmmss", null, 15),
+            DatePattern("yyyyMMdd'T'HHmmss'Z'", TimeZone.getTimeZone("GMT"), 16))
+
+        private const val FORMAT_DATE_PATTERN = "yyyyMMdd"
+
+        private val BYDAY_VALUES = arrayOf("SU", "MO", "TU", "WE", "TH", "FR", "SA")
+    }
+}
+
+/**
+ * Exception thrown by [RRuleFormatter.parse] when RRule couldn't be parsed.
+ */
+class RRuleParseException internal constructor(message: String, cause: Throwable? = null) :
+    IllegalArgumentException(message, cause)
+
+private fun parseError(message: String, cause: Throwable? = null): Nothing =
+    throw RRuleParseException(message, cause)
+
+private inline fun parseError(condition: Boolean, message: () -> String) {
+    contract {
+        returns() implies condition
+    }
+    if (!condition) {
+        throw RRuleParseException(message())
     }
 }