+++ lib/src/test/kotlin/com/maltaisn/recurpicker/TestDateExtensions.kt
@@ -20,12 +20,14 @@ import java.text.ParseException
 import java.text.SimpleDateFormat
 import java.util.TimeZone
 
-val patterns = listOf(
-    "yyyy-MM-dd'T'HH:mm:ss.SSSXXX" to TimeZone.getTimeZone("GMT"),
-    "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'" to TimeZone.getTimeZone("GMT"),
-    "yyyy-MM-dd'T'HH:mm:ss.SSS" to TimeZone.getDefault(),
-    "yyyy-MM-dd" to TimeZone.getDefault()
+val datePatterns = listOf(
+    DatePattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'", TimeZone.getTimeZone("GMT"), 24),
+    DatePattern("yyyy-MM-dd'T'HH:mm:ss.SSSXXX", null, null),
+    DatePattern("yyyy-MM-dd'T'HH:mm:ss.SSS", TimeZone.getDefault(), 23),
+    DatePattern("yyyy-MM-dd", TimeZone.getDefault(), 10)
 )
+
+data class DatePattern(val pattern: String, val timeZone: TimeZone?, val length: Int?)
 
 /**
  * Get UTC millis since epoch time for date patterns:
@@ -37,15 +39,17 @@ val patterns = listOf(
  */
 internal fun dateFor(date: String): Long {
     val dateFormat = SimpleDateFormat()
-    for ((pattern, timeZone) in patterns) {
-        if (timeZone != null) {
-            dateFormat.timeZone = timeZone
-        }
-        dateFormat.applyPattern(pattern)
-        return try {
-            dateFormat.parse(date)?.time ?: continue
-        } catch (e: ParseException) {
-            continue
+    for (pattern in datePatterns) {
+        if (pattern.length == null || date.length == pattern.length) {
+            if (pattern.timeZone != null) {
+                dateFormat.timeZone = pattern.timeZone
+            }
+            dateFormat.applyPattern(pattern.pattern)
+            return try {
+                dateFormat.parse(date)?.time
+            } catch (e: ParseException) {
+                null
+            } ?: continue
         }
     }
     throw IllegalArgumentException("Invalid date literal")